more build changes

+++ installer/scripts/build-natron.sh
@@ -4,63 +4,41 @@
 # Written by Ole-André Rodlie <olear@fxarena.net>
 #
 
-gcc -v
-sleep 5
+source $(pwd)/common.sh || exit 1
+NATRON_BRANCH=$1
+if [ "$NATRON_BRANCH" == "" ]; then
+  NATRON_BRANCH=master
+fi
 
-GIT_NATRON=https://github.com/MrKepzie/Natron.git
-
-if [ "$1" == "workshop" ];then
-NATRON_REL_V=$(cat tags/NATRON_WORKSHOP)
+if [ "$NATRON_BRANCH" == "workshop" ]; then
+  NATRON_REL_V=$NATRON_DEVEL_GIT
 else
-NATRON_REL_V=$(cat tags/NATRON_RELEASE)
+  NATRON_REL_V=$NATRON_STABLE_GIT
 fi
 
-if [ "$CUSTOM" != "" ]; then
-  NATRON_REL_V=$CUSTOM
+if [ "$NATRON_REL_V" == "" ]; then
+  echo "No git version defined, please check common.sh."
+  exit 1
 fi
 
-NATRON_REL_B=workshop
-
-if [ -z "$SDK_VERSION" ]; then
-  SDK_VERSION=2.0
-fi
-
-# Threads
-if [ -z "$MKJOBS" ]; then
-  MKJOBS=4
-fi
-
-# Setup
-if [ -z "$ARCH" ]; then
-  case "$( uname -m )" in
-    i?86) export ARCH=i686 ;;
-    amd64) export ARCH=x86_64 ;;
-       *) export ARCH=$( uname -m ) ;;
-  esac
-fi
-if [ "$ARCH" = "i686" ]; then
-  BF="-O2 -march=i686 -mtune=i686"
-elif [ "$ARCH" = "x86_64" ]; then
-  BF="-O2 -fPIC"
-else
-  BF="-O2"
-fi
-OS=$(uname -o)
-CWD=$(pwd)
-
-if [ "$OS" == "GNU/Linux" ]; then
-  INSTALL_PATH=/opt/Natron-$SDK_VERSION
-else
+if [ "$OS" != "GNU/Linux" ]; then
   INSTALL_PATH=/usr/local
+else
+  if [ ! -d $INSTALL_PATH ]; then
+    if [ -f $SRC_PATH/Natron-$SDK_VERSION-Linux-$ARCH-SDK.tar.xz ]; then
+      echo "Found binary SDK, extracting ..."
+      tar xvJf $SRC_PATH/Natron-$SDK_VERSION-Linux-$ARCH-SDK.tar.xz -C $SDK_PATH/ || exit 1
+    else
+      echo "Need to build SDK ..."
+      sh $CWD/installer/scripts/build-sdk.sh || exit 1
+    fi
+  fi
 fi
-
-TMP_PATH=$CWD/tmp
 
 if [ -d $TMP_PATH ]; then
   rm -rf $TMP_PATH || exit 1
 fi
 mkdir -p $TMP_PATH || exit 1
-mkdir -p $CWD/src
 
 # Setup env
 export PKG_CONFIG_PATH=$INSTALL_PATH/lib/pkgconfig:$INSTALL_PATH/libdata/pkgconfig
@@ -72,51 +50,43 @@ export PYTHON_HOME=$INSTALL_PATH
 export PYTHON_PATH=$INSTALL_PATH/lib/python3.4
 export PYTHON_INCLUDE=$INSTALL_PATH/include/python3.4
 
+echo
+echo "Building Natron $NATRON_REL_V from $NATRON_BRANCH against SDK $SDK_VERSION on $ARCH using $MKJOBS threads."
+echo
+sleep 5
+
 # Install natron
 cd $TMP_PATH || exit 1
 
-if [ -f $CWD/src/Natron-$NATRON_REL_V.tar.gz ]; then
-tar xvf $CWD/src/Natron-$NATRON_REL_V.tar.gz || exit 1
-cd Natron* || exit 1
+if [ -f $SRC_PATH/Natron-$NATRON_REL_V.tar.gz ]; then
+  tar xvf $SRC_PATH/Natron-$NATRON_REL_V.tar.gz || exit 1
+  cd Natron* || exit 1
 else
-git clone $GIT_NATRON || exit 1
-cd Natron || exit 1
-
-git checkout ${NATRON_REL_V} || exit 1
-#git checkout workshop || exit 1
-REL_GIT_VERSION=$(git log|head -1|awk '{print $2}')
-if [ "$NATRON_REL_V" != "$REL_GIT_VERSION" ]; then
-  echo "version mismatch: $NATRON_REL_V vs. $REL_GIT_VERSION"
-  exit 1
+  git clone $GIT_NATRON || exit 1
+  cd Natron || exit 1
+  git checkout $NATRON_REL_V || exit 1
+  REL_GIT_VERSION=$(git log|head -1|awk '{print $2}')
+  if [ "$NATRON_REL_V" != "$REL_GIT_VERSION" ]; then
+    echo "version mismatch: $NATRON_REL_V vs. $REL_GIT_VERSION"
+    exit 1
+  fi
+  git submodule update -i --recursive || exit 1
+  if [ "$NOSRC" != "1" ]; then
+    (cd .. ;
+      cp -a Natron Natron-$REL_GIT_VERSION
+      (cd Natron-$REL_GIT_VERSION ; find . -type d -name .git -exec rm -rf {} \;)
+      (cd Natron-$REL_GIT_VERSION/Gui/Resources/OpenColorIO-Configs ; find . -type d -name .git -exec rm -rf {} \;)
+      tar cvvzf $SRC_PATH/Natron-$REL_GIT_VERSION.tar.gz Natron-$REL_GIT_VERSION
+    )
+  fi
 fi
 
-git submodule update -i --recursive || exit 1
-
-(cd .. ;
-  cp -a Natron Natron-$REL_GIT_VERSION
-  (cd Natron-$REL_GIT_VERSION ; find . -type d -name .git -exec rm -rf {} \;)
-  tar cvvzf $CWD/src/Natron-$REL_GIT_VERSION.tar.gz Natron-$REL_GIT_VERSION
-)
-fi
-
-cat $CWD/installer/GitVersion.h | sed "s#__BRANCH__#${NATRON_REL_B}#;s#__COMMIT__#${REL_GIT_VERSION}#" > Global/GitVersion.h || exit 1
+cat $CWD/installer/natron/GitVersion.h | sed "s#__BRANCH__#${NATRON_BRANCH}#;s#__COMMIT__#${REL_GIT_VERSION}#" > Global/GitVersion.h || exit 1
 
 if [ "$OS" == "GNU/Linux" ]; then
-  cat $CWD/installer/config.pri > config.pri || exit 1
+  cat $CWD/installer/natron/config.pri > config.pri || exit 1
 else
-  cat $CWD/installer/config-freebsd.pri > config.pri || exit 1
-fi
-
-# fix for Linux
-if [ "$OS" == "GNU/Linux" ]; then
-  patch -p0< $CWD/patches/stylefix.diff || exit 1
-  #if [ "$1" == "workshop" ];then
-  #  patch -p0< $CWD/patches/gcc47fix.diff || exit 1
-  #fi
-else
-if [ "$1" == "workshop" ]; then
-patch -p0< $CWD/patches/freebsd-buildfix.diff || exit 1
-fi
+  cat $CWD/installer/natron/config-freebsd.pri > config.pri || exit 1
 fi
 
 rm -rf build
@@ -133,6 +103,9 @@ make -j${MKJOBS} || exit 1
 
 cp App/Natron $INSTALL_PATH/bin/ || exit 1
 cp Renderer/NatronRenderer $INSTALL_PATH/bin/ || exit 1
+if [ "$NATRON_BRANCH" == "workshop" ] && [ "$OS" == "GNU/Linux" ]; then
+  cp CrashReporter/NatronCrashReporter $INSTALL_PATH/bin/ || exit 1
+fi
 
 rm -rf * || exit 1
 
@@ -146,6 +119,9 @@ fi
 make -j${MKJOBS} || exit 1
 cp App/Natron $INSTALL_PATH/bin/Natron.debug || exit 1
 cp Renderer/NatronRenderer $INSTALL_PATH/bin/NatronRenderer.debug || exit 1
+if [ "$NATRON_BRANCH" == "workshop" ] && [ "$OS" == "GNU/Linux" ]; then
+  cp CrashReporter/NatronCrashReporter $INSTALL_PATH/bin/NatronCrashReporter.debug || exit 1
+fi
 fi
 
 cp -a ../Gui/Resources/OpenColorIO-Configs $INSTALL_PATH/share/ || exit 1
